Rearranged to allow for common JavaScript components.
It's a bit of a faff, but I patterned it after https://github.com/bradfordlemley/cra-monorepo-examples
This involved switching from npm to yarn (it just wouldn't work with npm), but I've wrapped everything
up in package.json, so you don't have to install yarn or lerna seperately and you can continue to use
npm as before.  Alternatively, the gradle build scripts will handle it all for you.